# openapi-spec.yaml
openapi: 3.0.3
info:
  title: 智慧乡村综合服务平台 API
  description: |
    智慧乡村综合服务平台的完整 API 规范，涵盖村民管理、村务治理、信息公示、生活服务四大核心模块。

    ## 架构说明
    - **主API服务器** (端口3001): 监控、稳定性管理、多语言支持、通知服务
    - **村务服务器** (端口5000): 核心村务管理、Socket.IO实时通信、紧急广播

    ## 核心功能模块
    1. **村委管理**: 村委信息管理、村务协同、任务调度
    2. **村民管理**: 信息数字化、在线办事、档案管理
    3. **村务治理**: 财务管理、项目管理、公开透明
    4. **信息公示**: 政策公告、村务通知、关键词查询
    5. **安全防护**: 数据加密、权限管控、操作审计
  version: 2.0.0
  contact:
    name: 智慧乡村平台团队
    email: support@smartvillage.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: 主API服务器 (监控、通知、多语言)
  - url: http://localhost:5000
    description: 村务服务器 (核心服务、Socket.IO)
...
paths:
...
  /api/residents:
    get:
      tags:
        - 村民管理
      summary: 获取村民档案列表
      description: 分页获取村民档案，支持多条件筛选和隐私保护
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          schema:
            type: string
          description: 姓名搜索（模糊匹配）
        - name: familyType
          in: query
          schema:
            type: string
            enum: [normal, poverty, single, disabled, veteran]
          description: 家庭类型筛选
        - name: ageRange
          in: query
          schema:
            type: string
            pattern: '^\d+-\d+$'
          description: 年龄范围（如：18-60）
      responses:
        '200':
          description: 村民档案列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ResidentProfile'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '403':
          description: 无权限访问

    post:
      tags:
        - 村民管理
      summary: 新建村民档案
      description: 创建新村民档案，自动生成一户一码
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResidentCreate'
      responses:
        '201':
          description: 档案创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/ResidentProfile'
                  householdQRCode:
                    type: string
                    description: 一户一码二维码URL
  /api/residents/{id}:
    get:
      tags:
        - 村民管理
      summary: 获取村民详细信息
      description: 根据ID获取村民详细档案，支持血缘关系验证
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: 村民ID
        - name: relation
          in: query
          schema:
            type: string
            enum: [self, spouse, parent, child, sibling]
          description: 查询关系（用于权限控制）
      responses:
        '200':
          description: 村民详细信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResidentDetail'
        '403':
          description: 无权限查看此信息
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - 村民管理
      summary: 更新村民信息
      description: 更新村民档案信息，记录变更日志
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResidentUpdate'
      responses:
        '200':
          description: 更新成功
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
  schemas:
...
    Pagination:
      type: object
      properties:
        current:
          type: integer
          description: 当前页码
        pageSize:
          type: integer
          description: 每页大小
        total:
          type: integer
          description: 总记录数
        totalPages:
          type: integer
          description: 总页数

    # ==================== 认证相关 ====================
...
    ResidentProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          description: 姓名
        phone:
          type: string
          description: 联系电话（脱敏）
        idCard:
          type: string
          description: 身份证号（脱敏）
        address:
          type: string
        familyType:
          type: string
          enum: [normal, poverty, single, disabled, veteran]
        householdSize:
          type: integer
        registrationDate:
          type: string
          format: date
        lastUpdated:
          type: string
          format: date-time
        qrCode:
          type: string
          description: 一户一码
    ResidentDetail:
      allOf:
        - $ref: '#/components/schemas/ResidentProfile'
        - type: object
          properties:
            fullIdCard:
              type: string
              description: 完整身份证号（需权限）
            fullPhone:
              type: string
              description: 完整手机号（需权限）
            familyMembers:
              type: array
              items:
                $ref: '#/components/schemas/FamilyMember'
            healthStatus:
              type: object
              properties:
                chronicDisease:
                  type: array
                  items:
                    type: string
                vaccinationStatus:
                  type: string
                medicalInsurance:
                  type: boolean
            economicStatus:
              type: object
              properties:
                income:
                  type: string
                  enum: [poverty, low, medium, high]
                subsidies:
                  type: array
                  items:
                    type: object
                    properties:
                      type:
                        type: string
                      amount:
                        type: number
                      startDate:
                        type: string
                        format: date
    ResidentCreate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 20
        phone:
          type: string
          pattern: '^1[3-9]\d{9}$'
        idCard:
          type: string
        address:
          type: string
        familyType:
          type: string
          enum: [normal, poverty, single, disabled, veteran]
        familyMembers:
          type: array
          items:
            $ref: '#/components/schemas/FamilyMemberCreate'
      required:
        - name
        - phone
        - idCard
        - address
    ResidentUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        familyType:
          type: string
        emergencyContact:
          type: string
        notes:
          type: string
    FamilyMember:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        relationship:
          type: string
          enum: [spouse, parent, child, sibling, grandparent, grandchild, other]
        age:
          type: integer
        idCard:
          type: string
        phone:
          type: string
        occupation:
          type: string
    FamilyMemberCreate:
      type: object
      properties:
        name:
          type: string
        relationship:
          type: string
          enum: [spouse, parent, child, sibling, grandparent, grandchild, other]
        age:
          type: integer
        idCard:
          type: string
        phone:
          type: string
        occupation:
          type: string
      required:
        - name
        - relationship
        - age

    # ==================== 村务治理 ====================
...
  responses:
...
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "权限不足"
              requiredPermission:
                type: string
    NotFound:
      description: 资源未找到
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "资源未找到"
              resource:
                type: string
...